[#15] go.mod: Update go version to 1.16
