+++ ecdsa.go
@@ -85,7 +85,7 @@ func MarshalPublicKey(key *ecdsa.PublicKey) []byte {
 		return nil
 	}
 
-	return encodePoint(key.X, key.Y)
+	return elliptic.MarshalCompressed(curve, key.X, key.Y)
 }
 
 // UnmarshalPublicKey from bytes.
@@ -99,6 +99,28 @@ func UnmarshalPublicKey(data []byte) *ecdsa.PublicKey {
 	}
 
 	return nil
+}
+
+func decodePoint(data []byte) (x *big.Int, y *big.Int) {
+	// empty data
+	if len(data) == 0 {
+		return
+	}
+
+	// tries to unmarshal compressed form
+	// returns (nil, nil) when:
+	// - wrong len(data)
+	// - data[0] != 2 && data[0] != 3
+	if x, y = elliptic.UnmarshalCompressed(curve, data); x != nil && y != nil {
+		return x, y
+	}
+
+	// tries to unmarshal uncompressed form and check that points on curve
+	if x, y = unmarshalXY(data); x == nil || y == nil || !curve.IsOnCurve(x, y) {
+		x, y = nil, nil
+	}
+
+	return x, y
 }
 
 // UnmarshalPrivateKey from bytes.
