+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/nspcc-dev/neofs-crypto
 
-go 1.13
+go 1.16
 
 require (
 	github.com/mr-tron/base58 v1.2.0
