+++ load.go
@@ -3,7 +3,7 @@ package crypto
 import (
 	"crypto/ecdsa"
 	"encoding/hex"
-	"io/ioutil"
+	"os"
 
 	"github.com/pkg/errors"
 )
@@ -13,7 +13,7 @@ import (
 // - hex string
 // - file path (D-bytes or SEC 1 / ASN.1 DER form)
 func LoadPrivateKey(val string) (*ecdsa.PrivateKey, error) {
-	if data, err := ioutil.ReadFile(val); err == nil {
+	if data, err := os.ReadFile(val); err == nil {
 		return UnmarshalPrivateKey(data)
 	} else if data, err = hex.DecodeString(val); err == nil {
 		return UnmarshalPrivateKey(data)
